refactor: improve serviceVariables hook
